Rename server -> core_server
